duplication settings

+++ src/services/apiCabins.js
@@ -39,10 +39,16 @@ export async function createEditCabin(newCabin, id) {
   }
 
   // 2. upload image
-  const { error: storageError } = await supabase.storage.from('cabin-images').upload(imageName, newCabin.image)
+  if(hasImagePath) return fromSnakeToCamel(data)
+     
+  const { error: storageError } = await supabase.storage
+    .from('cabin-images')
+    .upload(imageName, newCabin.image)
+
+  // 3. delete the cabin IF there was an error uploading image
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id)
-    console.log(storageError)
+    console.error(storageError)
     throw new Error('Cabins image could not be uploaded and the cabin was not created.')
   }
